+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.44.1...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- wrap function with functools.wraps
 
 [0.44.1](https://github.com/jshwi/docsig/releases/tag/v0.44.1) - 2023-12-01
 ------------------------------------------------------------------------