+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - exclude option
 
 ### Fixed
+- allow decorators to be referenced with module name
 - allow checking of files not ending in .py
 
 [0.44.3](https://github.com/jshwi/docsig/releases/tag/v0.44.3) - 2024-03-24